+++ tests/test_commands.py
@@ -531,6 +531,7 @@ def test_duplicate_option(invoke):
         "  -q, --quiet                  Decrease the default WARNING verbosity by one\n"
         "                               level for each additional repetition of the\n"
         "                               option.  [default: 0]\n"
+        "  --debug                      Shorthand for --verbosity DEBUG.\n"
         "  --tree                       Show the tree of nested subcommands and exit.\n"
         "  --man                        Read the command's manual page and exit.\n"
         "  --help-format [carapace|json|json-full|man|markdown|markdown-full]\n"